Full Update
Has Particle Effect for guns

Has Collision detection for picked up objects colliding with collidable objects. Unfortuantely, Level 1 is full of terrain and objects that are uncollidable. So this fix doesn't solve the chicken falling thru the world bug.

BUT I added a duct tape fix that basically teleports the chicken to y = 5, if they ever fall below y = -10. which should cover most of the village. Doesn't cover if the player carries the chicken up a hill and drop the chicken under the hill. The chicken teleport won't be high enough.